Update terminal.h
Kevin and I talked and he helped me figure these out better. So basically the terminal is just menus, and depending on the user's number and choices, then Data_center methods will be called through the data_link. We figured that the Data_center should have the provider directory since that is where the services are being added and where everything else is going on. Same reason why the pointers have also been moved to the Data_center. And then I added the functionality to the Data_center. The return types may need to be changed though. I'm not very good at coming up with meaningful ones. We thought that it wouldn't be a big deal if I just edited this branch.